Python/usedcarmart-backend/requirements.txt
Flask==2.2.3
supabase==0.2.4
pyjwt==2.7.0
python-dotenv==1.0.0
